64-bit compatibility changes for the sandbox code This CL contains some basic changes that eliminate some of the warnings that appear when the sandbox code is compiled for 64-bit Windows. This is part of a larger effort to support Native Client on 64-bit Windows (that will require the sandbox to support 64-bit Windows). TEST=will be tested when the rest of the code builds for 64-bit Windows BUG=27218 Review URL: http://codereview.chromium.org/378030
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@31625 0039d316-1c4b-4281-b951-d872f2087c98

+++ sandbox/src/service_resolver.cc
@@ -117,7 +117,7 @@ bool WriteProtectedChildMemory(HANDLE child_process,
                           PAGE_WRITECOPY, &old_protection))
     return false;
 
-  DWORD written;
+  SIZE_T written;
   bool ok = ::WriteProcessMemory(child_process, address, buffer, length,
                                  &written) && (length == written);
 
@@ -198,7 +198,7 @@ size_t ServiceResolverThunk::GetThunkSize() const {
 
 bool ServiceResolverThunk::IsFunctionAService(void* local_thunk) const {
   ServiceEntry function_code;
-  DWORD read;
+  SIZE_T read;
   if (!::ReadProcessMemory(process_, target_, &function_code,
                            sizeof(function_code), &read))
     return false;
@@ -294,7 +294,7 @@ NTSTATUS ServiceResolverThunk::PerformPatch(void* local_thunk,
   size_t thunk_size = GetThunkSize();
 
   // copy the local thunk buffer to the child
-  DWORD written;
+  SIZE_T written;
   if (!::WriteProcessMemory(process_, remote_thunk, local_thunk,
                             thunk_size, &written))
     return STATUS_UNSUCCESSFUL;
@@ -320,7 +320,7 @@ NTSTATUS ServiceResolverThunk::PerformPatch(void* local_thunk,
 bool ServiceResolverThunk::SaveOriginalFunction(void* local_thunk,
                                                 void* remote_thunk) {
   ServiceEntry function_code;
-  DWORD read;
+  SIZE_T read;
   if (!::ReadProcessMemory(process_, target_, &function_code,
                            sizeof(function_code), &read))
     return false;
@@ -355,7 +355,7 @@ bool ServiceResolverThunk::SaveOriginalFunction(void* local_thunk,
 
 bool Wow64ResolverThunk::IsFunctionAService(void* local_thunk) const {
   Wow64Entry function_code;
-  DWORD read;
+  SIZE_T read;
   if (!::ReadProcessMemory(process_, target_, &function_code,
                            sizeof(function_code), &read))
     return false;
@@ -381,7 +381,7 @@ bool Wow64ResolverThunk::IsFunctionAService(void* local_thunk) const {
 
 bool Win2kResolverThunk::IsFunctionAService(void* local_thunk) const {
   ServiceEntry function_code;
-  DWORD read;
+  SIZE_T read;
   if (!::ReadProcessMemory(process_, target_, &function_code,
                            sizeof(function_code), &read))
     return false;
